[refactor/#29] request 패키지 홈에 추가
